Move printing of misspelled word context to main
Move the printing of potentially misspelled word in context in their
spellchecked files to the main function. This make it easier to test the
script.

Also handle the case in the word_context function when there are no
lines to print for any reason, such as none of the specified words being
present in the content lines.