Delete the 'id' attribute from every HTML tag
All spellchecked files are converted to HTML and back again, so that
certain elements can be removed or adjusted before spellchecking. Pandoc
is used for this. Pandoc will add and 'id' tag to elements, which is
sometimes added in the final markdown.

Since we don't want to spellcheck automatically generated IDs, these
'id' attributes are now removed during preprocessing.